chore: 명료성을 위해 stock을 quantity 로 변경 패키지 명도 optionStock이 아닌 productStock까지 포괄하게 stock으로 변경
